Move ConsentBannerButton to a new "components" folder
